Bug fix: summary of summaries too long
If summarizing very long works, such as books, the summaries
taken together would exceed context length for the
final summary. Thus, if the summaries taken together
exceed 5000 words, it runs the summary chunking
process again. Also added code to save intermediary
summaries for later review. There is a potential risk
that even the chunked summary of summary of summaries
will exceed 8000 tokens. In that case there needs to be some
recursive function to fix it, and I am too busy today
to do it. I will add it later, however.